Make the caching code more forgiving
I updated last_modified in last commit, but last_modified saved in
browsers includes timezone info, which cause the page to not
load. One way to future proof against further changes to
last_modified is to ignore ALL errors (but requires dilligent log
checking so that the actual bugs are noticed!).