Add an alternative way for getting matches
It seems that most of the matches which display the "There are no stats
for this match" message only do so, because they call the internal
HiRez API with wrong "event ID". HiRez will probably eventually fix it,
but in the meantime I added a couple of scripts which I used to get the
data for the missing matches.