Add Version table to the database
To simplify future migrations. Also remove the code for the older
migrations, as it would be annoying to embed them in this new system.
And anyways, if they are still needed for some reason, they can be
looked up in source control.

Run create_db & migrate_db in the run_backend_debug.py script to
simplify local development.