Clean up fix_hirez.py 2/2
I wanted to keep all the applied fixes in the fix_hirez.py file, so
that the current database can be reconstructed with the updater logs
plus running fix_hirez.py, but the file is getting slightly unwieldy,
and also it isn't really needed to have the fixes under source control.
Hence from now on fix_hirez.py will be gitignored (but the helper
functions were moved to fix_hirez_lib.py).

Also, changes in the database will now be logged using sqldiff in
be/diffs.