Add alerter and refactor updater
Sends notifications when there are warnings or worse in logs.
Made a new script, so that I can control when it is launched, and
notifications will not be sent in the middle of the night.

Also changed updater a bit - added some prints for when it is run
manually, removed logging.basicConfig, logs moved to a subfolder and
named with datetime instead date + index and added env variable
MATCHES_WITH_NO_STATS - error will be logged if a match has no stats
and it is not specified in the env var.