Cascade deletion for build items
Slightly tricky in SQLAlchemy. In short (I think):
ondelete='CASCADE' causes it to add the FK constraint in the DDL.
passive_deletes=True means, that if some build_items are unloaded at the time
    of build deletion, then they will not be loaded and deleted (since they will
    be handled by the cascade).
cascade="all" causes it to invalidate the loaded build_items.

Also, cascade deletion works in SQLite only when the PRAGMA foreign_keys = ON;
option is used.